+++ app/src/main/java/com/studypulse/app/feat/semester/domain/SemesterSummaryRepository.kt
@@ -6,14 +6,14 @@ import kotlinx.coroutines.flow.Flow
 interface SemesterSummaryRepository {
     suspend fun put(minAttendance: Int): Result<Unit>
     suspend fun get(): Result<SemesterSummary>
-    suspend fun incPresent(by: Int): Result<Unit>
-    suspend fun decPresent(by: Int): Result<Unit>
-    suspend fun incAbsent(by: Int): Result<Unit>
-    suspend fun decAbsent(by: Int): Result<Unit>
-    suspend fun incUnmarked(by: Int): Result<Unit>
-    suspend fun decUnmarked(by: Int): Result<Unit>
-    suspend fun incCancelled(by: Int): Result<Unit>
-    suspend fun decCancelled(by: Int): Result<Unit>
+    suspend fun incPresent(by: Long): Result<Unit>
+    suspend fun decPresent(by: Long): Result<Unit>
+    suspend fun incAbsent(by: Long): Result<Unit>
+    suspend fun decAbsent(by: Long): Result<Unit>
+    suspend fun incUnmarked(by: Long): Result<Unit>
+    suspend fun decUnmarked(by: Long): Result<Unit>
+    suspend fun incCancelled(by: Long): Result<Unit>
+    suspend fun decCancelled(by: Long): Result<Unit>
 
      fun getSummaryFlow(): Flow<SemesterSummary>
 }
